Naive, untested improvement of AdvanceMap task
I'm not sure which of these tmClient functions is capable of actually throwing errors, but I generally like to give each potential source of error its own try-catch with more specific logging message.

DownloadRandomMap catches all of its own exceptions, so it shouldn't have any errors, but, it could potentially spin forever if it never succeeds, so I thought having a breakout case for lots of failures would be a good idea.

I didn't look at the four usages of AdvanceMap for if they are actually catching errors that might be getting thrown from this, but since it is already throwing errors, I decided to keep doing so here.